+++ src/build.task.ts
@@ -3,22 +3,27 @@ import type {Task} from './task/task.js';
 import {createBuild} from './project/build.js';
 import type {MapInputOptions, MapOutputOptions, MapWatchOptions} from './project/build.js';
 import {getDefaultEsbuildOptions} from './build/esbuildBuildHelpers.js';
-import {isThisProjectGro, toBuildOutPath} from './paths.js';
+import {isThisProjectGro, paths} from './paths.js';
 import {Timings} from './utils/time.js';
 import {loadGroConfig} from './config/config.js';
 import type {GroConfig} from './config/config.js';
 import {configureLogLevel} from './utils/log.js';
 import type {BuildConfig} from './config/buildConfig.js';
+import {buildSourceDirectory} from './build/buildSourceDirectory.js';
+import type {SpawnedProcess} from './utils/process.js';
+import type {TaskEvents as ServerTaskEvents} from './server.task.js';
+import {hasGroServer} from './config/defaultBuildConfig.js';
 
 export interface TaskArgs {
-	watch?: boolean;
 	mapInputOptions?: MapInputOptions;
 	mapOutputOptions?: MapOutputOptions;
 	mapWatchOptions?: MapWatchOptions;
+	closeApiServer?: (spawned: SpawnedProcess) => Promise<void>; // let other tasks hang onto the api server
 }
 
-export interface TaskEvents {
+export interface TaskEvents extends ServerTaskEvents {
 	'build.createConfig': (config: GroConfig) => void;
+	'build.prebuild': void;
 }
 
 export const task: Task<TaskArgs, TaskEvents> = {
@@ -38,7 +43,6 @@ export const task: Task<TaskArgs, TaskEvents> = {
 		if (dev) {
 			log.warn('building in development mode; normally this is only for diagnostics');
 		}
-		const watch = args.watch ?? false;
 		const {mapInputOptions, mapOutputOptions, mapWatchOptions} = args;
 
 		const timingToLoadConfig = timings.start('load config');
@@ -49,10 +53,30 @@ export const task: Task<TaskArgs, TaskEvents> = {
 
 		const esbuildOptions = getDefaultEsbuildOptions(config.target, config.sourcemap, dev);
 
+		// TODO think this through
+		// This is like a "prebuild" phase.
+		// Build everything with esbuild and Gro's `Filer` first,
+		// so we have the production server available to run while SvelteKit is building.
+		// See the other reference to `isThisProjectGro` for comments about its weirdness.
+		let spawnedApiServer: SpawnedProcess | null = null;
+		if (!isThisProjectGro) {
+			await buildSourceDirectory(config, dev, log);
+			events.emit('build.prebuild');
+
+			// now that the prebuild is ready, we can start the API server, if it exists
+			if (await hasGroServer()) {
+				events.once('server.spawn', (spawned) => {
+					spawnedApiServer = spawned;
+				});
+				await invokeTask('server');
+			}
+		}
+
 		// Not every build config is built for the final `dist/`!
 		// Only those that currently have `dist: true` are output.
 		// This allows a project's `src/gro.config.ts`
 		// to control the "last mile" each time `gro build` is run.
+		// TODO maybe assign these to the `config` above?
 		const buildConfigsToBuild = config.builds.filter((buildConfig) => buildConfig.dist);
 		// For each build config that has `dist: true`,
 		// infer which of the inputs are actual source files,
@@ -64,13 +88,11 @@ export const task: Task<TaskArgs, TaskEvents> = {
 				const inputFiles = await resolveInputFiles(buildConfig);
 				log.info(`building "${buildConfig.name}"`, inputFiles);
 				if (inputFiles.length) {
-					const outputDir = toBuildOutPath(dev, buildConfig.name);
 					const build = createBuild({
 						dev,
 						sourcemap: config.sourcemap,
 						inputFiles,
-						outputDir,
-						watch,
+						outputDir: paths.dist,
 						mapInputOptions,
 						mapOutputOptions,
 						mapWatchOptions,
@@ -83,7 +105,16 @@ export const task: Task<TaskArgs, TaskEvents> = {
 			}),
 		);
 
-		// ...
+		// done! clean up the API server
+		if (spawnedApiServer) {
+			if (args.closeApiServer) {
+				// don't await - whoever attached `closeApiServer` will clean it up
+				args.closeApiServer(spawnedApiServer);
+			} else {
+				spawnedApiServer!.child.kill();
+				await spawnedApiServer!.closed;
+			}
+		}
 	},
 };