+++ src/clean.task.ts
@@ -9,8 +9,7 @@ export interface TaskArgs {
 }
 
 export const task: Task<TaskArgs> = {
-	description:
-		'remove files: build/ (unless -B), dist/ (unless -D), and optionally .svelte/ (-s) and node_modules/ (-n)',
+	description: 'remove temporary dev and build files',
 	run: async ({log, args}): Promise<void> => {
 		// TODO document with mdsvex
 		await clean(