+++ src/version.task.ts
@@ -63,7 +63,7 @@ export const task: Task = {
 const confirmWithUser = async (versionIncrement: string, log: Logger): Promise<void> => {
 	const readline = createReadlineInterface({input: process.stdin, output: process.stdout});
 	log.info(green(versionIncrement), '← new version');
-	await new Promise(async (resolve) => {
+	await new Promise<void>(async (resolve) => {
 		const [latestChangelogVersion, currentPackageVersion] = await Promise.all([
 			getLatestChangelogHeading(),
 			getCurrentPackageVersion(),
@@ -81,7 +81,7 @@ const confirmWithUser = async (versionIncrement: string, log: Logger): Promise<v
 			}
 			log.info(rainbow('proceeding'));
 			readline.close();
-			resolve(null);
+			resolve();
 		});
 	});
 };