+++ changelog.md
@@ -1,5 +1,12 @@
 # Gro changelog
 
+## 0.11.4
+
+- integrate the default Gro API server with SvelteKit and Gro's build task
+  ([#145](https://github.com/feltcoop/gro/pull/145))
+- add task `gro server` to support the default API server use case in SvelteKit
+  ([#145](https://github.com/feltcoop/gro/pull/145))
+
 ## 0.11.3
 
 - add `events` to `TaskContext` and its generic type to `Task`,